split Passes.scala into multiple files(InferTypes.scala, Resolves.scala, RemoveCHIRRTL.scala)
